Re-organize imports and address flake8 warnings
