+++ auto_period_finder/__init__.py
@@ -1,2 +1,1 @@
-from auto_period_finder.enums import *
-from auto_period_finder.finders import *
+from .finders import *
