+++ auto_period_finder/enums/__init__.py
@@ -1,1 +1,3 @@
-from auto_period_finder.enums._enums import *
+from ._enums import TimeSeriesDecomposer
+
+__all__ = ["TimeSeriesDecomposer"]
